requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.17.0
pdfplumber==0.9.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
reportlab==4.0.4
matplotlib==3.7.2